requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
backtrader==1.9.76.123
ccxt==1.88.26
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.3
dateparser==1.1.1
frozenlist==1.3.0
idna==3.3
joblib==1.1.0
multidict==6.0.2
numpy==1.22.4
pandas==1.4.2
pycares==4.2.0
pycparser==2.21
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.0
schedule==1.1.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
ta==0.10.1
TA-Lib @ file:///C:/Users/Tammi/Dropbox/KustomDevOps/CloudApps/11_cryptosphere_v2/downloads/TA_Lib-0.4.24-cp39-cp39-win_amd64.whl
threadpoolctl==3.1.0
tzdata==2022.1
tzlocal==4.2
ujson==5.3.0
urllib3==1.26.9
websocket-client==1.3.3
websockets==10.3
yarl==1.7.2
